// backend/src/modules/goods-issue-note/enums/goods-issue-note-status.enum.ts
export enum GOODS_ISSUE_NOTE_STATUS {
  Nonexistent = 'goodsIssueNote.status.non_existent',
  Expired = 'goodsIssueNote.status.expired',
  Draft = 'goodsIssueNote.status.draft',
  Created = 'goodsIssueNote.status.created',
  Validated = 'goodsIssueNote.status.validated',
  Sent = 'goodsIssueNote.status.sent',
  Issued = 'goodsIssueNote.status.issued',
  Accepted = 'goodsIssueNote.status.accepted',
  Rejected = 'goodsIssueNote.status.rejected',
  Cancelled = 'goodsIssueNote.status.cancelled',
  Invoiced = 'goodsIssueNote.status.invoiced',
}