// backend/src/modules/quotation/enums/quotation-status.enum.ts
export enum QUOTATION_STATUS {
  Nonexistent = 'quotation.status.non_existent',
  Expired = 'quotation.status.expired',
  Draft = 'quotation.status.draft',
  Created = 'quotation.status.created',
  Validated = 'quotation.status.validated',
  Sent = 'quotation.status.sent',
  Accepted = 'quotation.status.accepted',
  Rejected = 'quotation.status.rejected',
  Invoiced = 'quotation.status.invoiced',
  Ordered = 'quotation.status.ordered',
  Delivered = 'quotation.status.delivered',
  Archived = 'quotation.status.archived',
}