// backend/src/modules/return-note/enums/return-note-status.enum.ts
export enum RETURN_NOTE_STATUS {
  Nonexistent = 'returnNote.status.non_existent',
  Expired = 'returnNote.status.expired',
  Draft = 'returnNote.status.draft',
  Validated = 'returnNote.status.validated',
  Sent = 'returnNote.status.sent',
  Accepted = 'returnNote.status.accepted',
  Rejected = 'returnNote.status.rejected',
  Cancelled = 'returnNote.status.cancelled',
  Invoiced = 'returnNote.status.invoiced',
  Archived = 'returnNote.status.archived',
}